// src/api/onlyfit-api.gen/models/SystemStatusGetResponseDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OnlyFit API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { MinimumVersions } from './MinimumVersions';
import {
    MinimumVersionsFromJSON,
    MinimumVersionsFromJSONTyped,
    MinimumVersionsToJSON,
    MinimumVersionsToJSONTyped,
} from './MinimumVersions';

/**
 *
 * @export
 * @interface SystemStatusGetResponseDTO
 */
export interface SystemStatusGetResponseDTO {
    /**
     *
     */
    availability?: SystemStatusGetResponseDTOAvailabilityEnum;
    /**
     *
     */
    retryAfterSeconds?: number;
    /**
     *
     */
    minimumVersions?: MinimumVersions;
    /**
     *
     */
    domainRouting?: { [key: string]: SystemStatusGetResponseDTODomainRoutingEnum; };
}


/**
 * @export
 */
export const SystemStatusGetResponseDTOAvailabilityEnum = {
    Available: 'AVAILABLE',
    Maintenance: 'MAINTENANCE',
} as const;
export type SystemStatusGetResponseDTOAvailabilityEnum = typeof SystemStatusGetResponseDTOAvailabilityEnum[keyof typeof SystemStatusGetResponseDTOAvailabilityEnum];

/**
 * @export
 */
export const SystemStatusGetResponseDTODomainRoutingEnum = {
    OnlyfitApi: 'ONLYFIT_API',
    OnlyfitCore: 'ONLYFIT_CORE',
} as const;
export type SystemStatusGetResponseDTODomainRoutingEnum = typeof SystemStatusGetResponseDTODomainRoutingEnum[keyof typeof SystemStatusGetResponseDTODomainRoutingEnum];


/**
 * Check if a given object implements the SystemStatusGetResponseDTO interface.
 */
export function instanceOfSystemStatusGetResponseDTO(value: object): value is SystemStatusGetResponseDTO {
    return true;
}

export function SystemStatusGetResponseDTOFromJSON(json: any): SystemStatusGetResponseDTO {
    return SystemStatusGetResponseDTOFromJSONTyped(json, false);
}

export function SystemStatusGetResponseDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): SystemStatusGetResponseDTO {
    if (json == null) {
        return json;
    }
    return {

        'availability': json['availability'] == null ? undefined : json['availability'],
        'retryAfterSeconds': json['retryAfterSeconds'] == null ? undefined : json['retryAfterSeconds'],
        'minimumVersions': json['minimumVersions'] == null ? undefined : MinimumVersionsFromJSON(json['minimumVersions']),
        'domainRouting': json['domainRouting'] == null ? undefined : json['domainRouting'],
    };
}

export function SystemStatusGetResponseDTOToJSON(json: any): SystemStatusGetResponseDTO {
    return SystemStatusGetResponseDTOToJSONTyped(json, false);
}

export function SystemStatusGetResponseDTOToJSONTyped(value?: SystemStatusGetResponseDTO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'availability': value['availability'],
        'retryAfterSeconds': value['retryAfterSeconds'],
        'minimumVersions': MinimumVersionsToJSON(value['minimumVersions']),
        'domainRouting': value['domainRouting'],
    };
}